Suggest-Service: Move classes into suggest package
