+++ suggest-service-parent/suggest-service/src/main/java/de/cxp/ocs/suggest/spi/SuggestServiceConfigProvider.java
@@ -1,0 +1,6 @@
+package de.cxp.ocs.spi;
+
+
+public interface SuggestServiceConfigProvider {
+
+}
